Move layer base classes into their own file
